Move StringIndex into it's own method
We need the IndexedModel, in the performMachineLearning function, so I had to elevate it up
into the static main method, then inject it into the static methods that need it. Really only 2 and the latter
is the whole restructuring change. I can now continue working on performing the machine learning piece.